+++ com.unity.render-pipelines.high-definition/Editor/ShaderGraph/Nodes/ParallaxOcclusionMappingNode.cs
@@ -11,7 +11,7 @@ using UnityEngine.Rendering;
 namespace UnityEditor.Experimental.Rendering.HDPipeline
 {
     [Title("Utility", "High Definition Render Pipeline", "Parallax Occlusion Mapping")]
-    class ParallaxOcclusionMappingNode : AbstractMaterialNode, IGeneratesBodyCode, IGeneratesFunction, IMayRequireViewDirection
+    class ParallaxOcclusionMappingNode : AbstractMaterialNode, IGeneratesBodyCode, IGeneratesFunction, IMayRequireViewDirection, IMayRequireMeshUV
     {
         public ParallaxOcclusionMappingNode()
         {
@@ -56,7 +56,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             AddSlot(new SamplerStateMaterialSlot(kHeightmapSamplerSlotId, kHeightmapSamplerSlotName, kHeightmapSamplerSlotName, SlotType.Input));
             AddSlot(new Vector1MaterialSlot(kAmplitudeSlotId, kAmplitudeSlotName, kAmplitudeSlotName, SlotType.Input, 1.0f, ShaderStageCapability.Fragment));
             AddSlot(new Vector1MaterialSlot(kStepsSlotId, kStepsSlotName, kStepsSlotName, SlotType.Input, 5.0f, ShaderStageCapability.Fragment));
-            AddSlot(new Vector2MaterialSlot(kUVsSlotId, kUVsSlotName, kUVsSlotName, SlotType.Input, Vector2.zero, ShaderStageCapability.Fragment));
+            AddSlot(new UVMaterialSlot(kUVsSlotId, kUVsSlotName, kUVsSlotName, UVChannel.UV0, ShaderStageCapability.Fragment));
             AddSlot(new Vector1MaterialSlot(kLodSlotId, kLodSlotName, kLodSlotName, SlotType.Input, 0.0f, ShaderStageCapability.Fragment));
             AddSlot(new Vector1MaterialSlot(kLodThresholdSlotId, kLodThresholdSlotName, kLodThresholdSlotName, SlotType.Input, 0.0f, ShaderStageCapability.Fragment));
 
@@ -152,8 +152,8 @@ $precision {6} = {3} * 0.01;
 // Transform the view vector into the UV space.
 $precision3 {7}    = normalize($precision3({4}.xy * {6}, {4}.z)); // TODO: skip normalize
 
-PerPixelHeightDisplacementParam {1};
-{1}.uv = {2};",
+PerPixelHeightDisplacementParam {0};
+{0}.uv = {1};",
                 tmpPOMParam,
                 uvs,
                 CoordinateSpace.Tangent.ToVariableName(InterpolatorType.ViewDirection),
@@ -187,6 +187,17 @@ $precision {6} = ({7} - {10} * {7}) / max({9}, 0.0001);
         public NeededCoordinateSpace RequiresViewDirection(ShaderStageCapability stageCapability = ShaderStageCapability.All)
         {
             return NeededCoordinateSpace.Tangent;
+        }
+
+        public bool RequiresMeshUV(UVChannel channel, ShaderStageCapability stageCapability)
+        {
+            if (channel != UVChannel.UV0)
+                return false;
+
+            if (IsSlotConnected(kUVsSlotId))
+                return false;
+
+            return true;
         }
     }
 }